backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
click==8.3.1
exceptiongroup==1.3.1
fastapi==0.128.2
h11==0.16.0
idna==3.11
pydantic==2.12.5
pydantic_core==2.41.5
pydantic_settings==2.12.0
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0
sqlalchemy
psycopg2-binary
python-dotenv
alembic
email-validator
httpx
ipdb
python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart
langchain
langchain-core
langchain-community
sentence-transformers
faiss-cpu
groq
langchain-groq
redis